+++ src/CodeCreator.php
@@ -40,47 +40,15 @@ class CodeCreator
         if (!isset($schema->properties)) {
             return [];
         }
-        $classes = [];
-        $namespace = new PhpNamespace($this->defaultNamespace);
-        $class = $namespace->addClass($this->defaultClass)
-                        ->addImplement('\JsonSerializable');
-        $constructor = $class->addMethod('__construct');
-        $constructorComment = [];
-        $constructorBody = '';
-        $serializableRequiredProperties = '';
-        $serializableOptionalProperties = '';
-        foreach ($schema->properties as $propertyName => $propertyAttributes) {
-            $property = $this->properties->create($propertyName, $propertyAttributes, $this->defaultClass, $this->defaultNamespace);
-            if ($this->isRequired($propertyName, $schema)) {
-                $constructorBody.= $property->constructorBody();
-                $constructorComment[] = $property->constructorComment();
-                $constructor = $property->addParameterTo($constructor);
-                $class = $property->addTo($class);
-                $serializableRequiredProperties.= $property->serializingCode();
-                $classes = array_merge($classes, $property->extraClasses($this));
+        $classes = $this->createClass($schema, $this->defaultClass);
+        $references = isset($schema->definitions) ? $schema->definitions : [];
+        foreach ($references as $className => $classDefinition) {
+            if (isset($classDefinition->type) && $classDefinition->type === 'object') {
+                $classes = array_merge($classes, $this->createClass($classDefinition, $className));
             } else {
-                $class = $property->addTo($class);
-                $property->addSetterTo($class);
-                $serializableOptionalProperties.= $property->optionalSerializingCode();
+                $classes[$className] = $this->createEnum($className, $classDefinition->enum);
             }
         }
-        $constructorComment = array_filter($constructorComment, function ($comment) {
-            return !empty($comment);
-        });
-        if (!empty($constructorComment)) {
-            $constructor->addComment(implode("\n", $constructorComment));
-        }
-        $constructor->addBody($constructorBody);
-        if (!empty($serializableOptionalProperties)) {
-            $serializableMethodBody = "\$values = [\n" . $serializableRequiredProperties . "];\n";
-            $serializableMethodBody.= $serializableOptionalProperties;
-            $serializableMethodBody.= "return \$values;\n";
-        } else {
-            $serializableMethodBody = "return [\n" . $serializableRequiredProperties . "];";
-        }
-        $class->addMethod('jsonSerialize')
-            ->addBody($serializableMethodBody);
-        $classes[$this->defaultClass] = $namespace;
         return $classes;
     }
 
@@ -112,11 +80,12 @@ class CodeCreator
         $constants = array_map(function ($constant) {
             return 'self::' . strtoupper($constant);
         }, $values);
-        $constructorBody = ' $possibleValues = [' . implode(', ', $constants) . '];
-            if (!in_array($value, $possibleValues)) {
-                throw new InvalidValueException($value . \' is not an allowed value for EnumPropertyFoo\');
-            }
-            $this->value = $value;';
+        $constantsList = implode(', ', $constants);
+        $constructorBody = " \$possibleValues = [$constantsList];\n" .
+            "if (!in_array(\$value, \$possibleValues)) {\n" .
+            "   throw new InvalidValueException(\$value . ' is not an allowed value for $className');\n" .
+            "}\n" .
+            "\$this->value = \$value;";
 
         $constructor = $class->addMethod('__construct')
                             ->addComment($constructorComment)
@@ -137,5 +106,56 @@ class CodeCreator
         $namespace->addClass($className)
             ->addExtend('\Exception');
         return $namespace;
+    }
+
+    /**
+     * @param mixed $schema
+     * @param string $className
+     * @return array
+     */
+    private function createClass($schema, $className)
+    {
+        $classes = [];
+        $namespace = new PhpNamespace($this->defaultNamespace);
+        $class = $namespace->addClass($className)
+            ->addImplement('\JsonSerializable');
+        $constructor = $class->addMethod('__construct');
+        $constructorComment = [];
+        $constructorBody = '';
+        $serializableRequiredProperties = '';
+        $serializableOptionalProperties = '';
+        foreach ($schema->properties as $propertyName => $propertyAttributes) {
+            $property = $this->properties->create($propertyName, $propertyAttributes, $className, $this->defaultNamespace);
+            if ($this->isRequired($propertyName, $schema)) {
+                $constructorBody .= $property->constructorBody();
+                $constructorComment[] = $property->constructorComment();
+                $constructor = $property->addParameterTo($constructor);
+                $class = $property->addTo($class);
+                $serializableRequiredProperties .= $property->serializingCode();
+                $classes = array_merge($classes, $property->extraClasses($this));
+            } else {
+                $class = $property->addTo($class);
+                $property->addSetterTo($class);
+                $serializableOptionalProperties .= $property->optionalSerializingCode();
+            }
+        }
+        $constructorComment = array_filter($constructorComment, function ($comment) {
+            return !empty($comment);
+        });
+        if (!empty($constructorComment)) {
+            $constructor->addComment(implode("\n", $constructorComment));
+        }
+        $constructor->addBody($constructorBody);
+        if (!empty($serializableOptionalProperties)) {
+            $serializableMethodBody = "\$values = [\n" . $serializableRequiredProperties . "];\n";
+            $serializableMethodBody .= $serializableOptionalProperties;
+            $serializableMethodBody .= "return \$values;\n";
+        } else {
+            $serializableMethodBody = "return [\n" . $serializableRequiredProperties . "];";
+        }
+        $class->addMethod('jsonSerialize')
+            ->addBody($serializableMethodBody);
+        $classes[$className] = $namespace;
+        return $classes;
     }
 }